+++ scripts/winesapos-upgrade.sh
@@ -680,6 +680,11 @@ if [[ "$(sudo cat /etc/winesapos/IMAGE_TYPE)" != "minimal" ]]; then
         echo "Adding Open Gamepad UI complete."
     fi
 fi
+
+${CMD_PACMAN} -Q jfsutils
+if [ $? -ne 0 ]; then
+    ${CMD_PACMAN_INSTALL} jfsutils
+fi
 echo "Running 3.3.0 to 3.4.0 upgrades complete."
 
 echo "Upgrading system packages..."