+++ scripts/repo/winesapos-build-repo.sh
@@ -52,6 +52,7 @@ makepkg_fn mbpfan-git
 makepkg_fn mesa-steamos
 makepkg_fn lib32-mesa-steamos
 makepkg_fn oh-my-zsh-git
+makepkg_fn oversteer
 makepkg_fn paru
 makepkg_fn plasma5-themes-vapos-steamos
 makepkg_fn qdirstat