+++ scripts/winesapos-install.sh
@@ -467,12 +467,11 @@ options radeon cik_support=0
 options amdgpu si_support=1
 options amdgpu cik_support=1" > ${WINESAPOS_INSTALL_DIR}/etc/modprobe.d/winesapos-amd.conf
 
-# Workaround known AMD driver issues.
+# Workaround a known AMD driver issue:
 # https://www.phoronix.com/news/AMDGPU-APU-noretry
-# https://gitlab.freedesktop.org/drm/amd/-/issues/2354
+# https://gitlab.freedesktop.org/drm/amd/-/issues/934
 echo "
-options amdgpu noretry=0
-options amdgpu sg_display=0" >> ${WINESAPOS_INSTALL_DIR}/etc/modprobe.d/winesapos-amd.conf
+options amdgpu noretry=0" >> ${WINESAPOS_INSTALL_DIR}/etc/modprobe.d/winesapos-amd.conf
 
 # Flatpak.
 pacman_install_chroot flatpak