+++ scripts/winesapos-upgrade.sh
@@ -808,6 +808,10 @@ fi
 sudo -E -u ${WINESAPOS_USER_NAME} ${qdbus_cmd} ${kdialog_dbus} /ProgressDialog org.kde.kdialog.ProgressDialog.close
 echo "Running 3.3.0 to 3.4.0 upgrades complete."
 
+echo "Running 3.4.0 to 4.0.0 upgrades..."
+sed -i s'/options amdgpu sg_display=0//'g /etc/modprobe.d/winesapos-amd.conf
+echo "Running 3.4.0 to 4.0.0 upgrades complete."
+
 echo "Upgrading system packages..."
 kdialog_dbus=$(sudo -E -u ${WINESAPOS_USER_NAME} kdialog --title "winesapOS Upgrade" --progressbar "Please wait for all system packages to upgrade (this can take a long time)..." 9 | cut -d" " -f1)
 sudo -E -u ${WINESAPOS_USER_NAME} ${qdbus_cmd} ${kdialog_dbus} /ProgressDialog showCancelButton false