+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 - Change ClamTk and ClamAV to be installed as native packages for stability reasons.
 - Change extra Linux firmware to be installed into all release images.
 - Change torrent utility from Transmission to qBittorrent.
-- Change Wine GE version from Wine-GE-Proton7-43 to Wine-GE-Proton8-17.
+- Change Wine GE version from Wine-GE-Proton7-43 to Wine-GE-Proton8-22.
 - Change GE Proton version from GE-Proton7-55 to GE-Proton8-22.
 - Add the Nix package manager.
 - Add a Gamescope session for Steam.