+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Add support for Razer accessories.
 - Change the 'vapor-steamos-theme-kde' package to be the newly renamed 'plasma5-themes-vapor-steamos' package.
 - Add Oversteer for managing racing wheels and related accessories.
+- Deprecate SteamOS packages (linux-steamos and mesa-steamos).
 
 ## 3.3.0 - The Major Upgrade Update