+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Change the 'vapor-steamos-theme-kde' package to be the newly renamed 'plasma5-themes-vapor-steamos' package.
 - Add Oversteer for managing racing wheels and related accessories.
 - Deprecate SteamOS packages (linux-steamos and mesa-steamos).
+- Add DOSBox for 16-bit Windows application support in Wine.
 
 ## 3.3.0 - The Major Upgrade Update